requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-RESTful==0.3.0
Flask-Testing==0.4.2
flask-script==2.0.3
Flask-OAuth==0.12
Flask-Classy==0.6.10
decorator==3.4.0
blinker==1.3
bugsnag==2.2.0
newrelic==2.46.0.37
gunicorn==19.3.0
pymongo==2.7.2
httplib2==0.9.1
nose==1.3.3
mock==1.0.1
requests==2.7.0
flask-redis==0.1.0
vaderSentiment==0.5
Flask-SQLAlchemy==2.0
stripe==1.23.0
tweepy==3.5.0
APScheduler==3.0.0
python-dateutil==2.5.0
ipdb==0.8.1
flask-sslify==0.1.5
git+https://github.com/dgrtwo/ParsePy.git